requirements.txt
Django==1.8.0
django-extensions==1.5.5
djangorestframework==3.1.3
psycopg2==2.6.1
